cambio: se movio de lugar la opcion para agregar una tarea
